requirements.txt
asgiref==3.5.0
blinker==1.4
Brlapi==0.8.2
certifi==2020.12.5
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.5.0
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
drf-excel==2.1.0
et-xmlfile==1.1.0
FormEncode==2.0.1
future==0.18.2
hidpidaemon==18.4.6
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
jeepney==0.7.1
keyring==23.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
louis==3.18.0
macaroonbakery==1.3.1
matplotlib==3.3.3
more-itertools==4.2.0
mysqlclient==2.1.0
netifaces==0.10.9
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.9
Pillow==9.0.1
pop-transition==1.1.2
protobuf==3.12.4
pycairo==1.16.2
pycups==2.0.1
pydbus==0.6.0
PyDispatcher==2.0.5
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt===2.2.1pop0-1628521831-21.10-5dbe6cf
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-dotenv==0.19.2
python-twitter==3.5
python-xlib==0.29
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
repolib==1.5.2
repoman==1.4.0
requests==2.25.1
requests-oauthlib==1.3.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
simplejson==3.17.2
six==1.15.0
SQLObject==3.9.1
sqlparse==0.4.2
systemd-python==234
tweet-preprocessor==0.6.0
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.2
wadllib==1.3.5
wordcloud==1.8.1
xdg==5
xkit==0.0.0
zipp==1.0.0
